+++ src/main/java/woowacourse/shoppingcart/dto/response/AddressResponse.java
@@ -1,4 +1,4 @@
-package woowacourse.shoppingcart.dto;
+package woowacourse.shoppingcart.dto.response;
 
 public class AddressResponse {
