+++ src/main/java/woowacourse/shoppingcart/dto/response/CustomerResponse.java
@@ -1,6 +1,7 @@
-package woowacourse.shoppingcart.dto;
+package woowacourse.shoppingcart.dto.response;
 
 import woowacourse.shoppingcart.domain.customer.Customer;
+import woowacourse.shoppingcart.dto.request.CustomerRequest;
 
 public class CustomerResponse {
