+++ src/main/java/woowacourse/shoppingcart/ui/CustomerController.java
@@ -18,9 +18,9 @@ import woowacourse.shoppingcart.application.CustomerService;
 import woowacourse.shoppingcart.application.dto.CustomerDto;
 import woowacourse.shoppingcart.application.dto.ModifiedCustomerDto;
 import woowacourse.shoppingcart.application.dto.SignInDto;
-import woowacourse.shoppingcart.dto.CustomerResponse;
-import woowacourse.shoppingcart.dto.ModifiedCustomerRequest;
-import woowacourse.shoppingcart.dto.SignUpRequest;
+import woowacourse.shoppingcart.dto.response.CustomerResponse;
+import woowacourse.shoppingcart.dto.request.ModifiedCustomerRequest;
+import woowacourse.shoppingcart.dto.request.SignUpRequest;
 
 @RestController
 @RequestMapping("/api")
